added verifier crate
