+++ distaff/Cargo.toml
@@ -14,12 +14,15 @@ path = "src/lib.rs"
 bench = false
 
 [features]
-concurrent = ["winterfell/concurrent"]
+concurrent = ["prover/concurrent", "std"]
+default = ["std"]
+std = ["air/std", "assembly/std", "processor/std", "prover/std", "verifier/std"]
 
 [dependencies]
-air = { package = "distaff-air", path = "../air" }
-assembly = { package = "distaff-assembly", path = "../assembly" }
-processor = { package = "distaff-processor", path = "../processor" }
+air = { package = "distaff-air", path = "../air", default-features = false }
+assembly = { package = "distaff-assembly", path = "../assembly", default-features = false }
+processor = { package = "distaff-processor", path = "../processor", default-features = false }
+prover = { package = "winter-prover", version = "0.2", default-features = false }
 hex = "0.4"
 log = "0.4"
-winterfell = "0.2"
+verifier = { package = "distaff-verifier", path = "../verifier", default-features = false }
